requirements.txt
flask
python-dotenv
requests
beautifulsoup4
lxml
schedule

